requirements.txt
langchain_core==0.3.75
langchain_openai==0.3.32
python-dotenv==1.1.1
streamlit==1.22.0
transformers==4.55.2
